feat: parallel batch [T20260509-2]
Tasks:
- T20260509-2: Modularize crates/orbit-engine/src/activity_job/job_executor.rs (2,841 lines)
